+++ src/spyglass/position/v1/position_dlc_project.py
@@ -14,10 +14,13 @@ import pandas as pd
 import ruamel.yaml
 
 from spyglass.common.common_lab import LabTeam
+from spyglass.position.v1.dlc_utils import (
+    _set_permissions,
+    check_videofile,
+    get_video_path,
+)
 from spyglass.settings import dlc_project_dir, dlc_video_dir
 from spyglass.utils.dj_mixin import SpyglassMixin
-
-from .dlc_utils import _set_permissions, check_videofile, get_video_path
 
 schema = dj.schema("position_v1_dlc_project")
 
@@ -491,7 +494,7 @@ class DLCProject(SpyglassMixin, dj.Manual):
                 f"{import_labeled_data_path.as_posix()}/{video_file}/*.h5"
             )[0]
             dlc_df = pd.read_hdf(h5_file)
-            dlc_df.columns.set_levels([team_name], level=0, inplace=True)
+            dlc_df.columns = dlc_df.columns.set_levels([team_name], level=0)
             dlc_df.to_hdf(
                 Path(
                     f"{current_labeled_data_path.as_posix()}/{video_file}/CollectedData_{team_name}.h5"