+++ src/spyglass/sharing/__init__.py
@@ -1,4 +1,4 @@
-from .sharing_kachery import (
+from spyglass.sharing.sharing_kachery import (
     AnalysisNwbfileKachery,
     AnalysisNwbfileKacherySelection,
     KacheryZone,